Fix OBI-Impulse strategy bugs and retune parameters
- Fix freshness gate: use wall-clock comparisons to prevent exchange timestamp drift
- Fix combination logic: HIGH conviction requires side match only, not venue match
- Fix signal expiry: use wall-clock stored_at_ns for reliable timeout in simulation
- Add momentum filter: impulse must agree with previous delta sign
- Retune parameters: impulse_threshold=10bps, entry_threshold=5bps, obi_persist=30ms, signal_timeout=250ms, lag_threshold=1.0bps
- Change lag_threshold_bps to f64 in schema for precision
- Update documentation: HANDOVER.md v0.1.4, architecture.md updates, new module_signal.md and environment_setup.md
- All 69 tests pass

+++ src/config/schema.rs
@@ -165,9 +165,9 @@ pub struct StrategySettings {
     #[validate(range(min = 1, max = 100))]
     pub impulse_threshold_bps: u64,
 
-    /// Max move on other exchange to consider it "lagging" (1-2 bps)
-    #[validate(range(min = 1, max = 50))]
-    pub lag_threshold_bps: u64,
+    /// Max move on other exchange to consider it "lagging" (0.5-5 bps can be fractional)
+    #[validate(range(min = 0.1, max = 50.0))]
+    pub lag_threshold_bps: f64,
 
     /// Lookback window in ms for price change detection
     #[validate(range(min = 1, max = 100))]
